Implemented procedures to remove duplicate proto-abstract values.  This prevents subset? from grinding to a halt due to these duplicates.
Special care had to be taken to prevent equal? from being applied to
identical primitive-procedures, which are circular and thus cause equal? to
yield a seg-fault.  A new procedure, equalq?, first tests if objects are eq?
before proceeding to look in the lists/vectors/structures.

There is still a seg-fault when running it in 0-CFA on saddle.vlad.  Don't
know why...