Native pairs added to flow analysis.
This is the version producing the results sent to Jeff on 08/31/06.  Some
cleanup is needed--taking out the old code made obsolete by adding the new
code included to do width and depth limits for native pairs.  This new code
actually is a more-general case and manages width limits for closures and
moving subtrees up in both width and depth limiting.

Short summary--
- 0cfa works for factorial, even-odd, foo (and subvariants of foo)
- precise tested on factorial and even-odd
- in-between completes on even-odd; not tested for other cases.  Correctness
  of result to be verified.