We used to have explicit functions to compute transform:   perturbation-transform   forward-transform   sensitivity-tranform   reverseify-parameter   reverse-transform and their inverses:   perturbation-transform-inverse   forward-transform-inverse   sensitivity-tranform-inverse   unreverseify-parameter   reverse-transform-inverse
We desire it to be the case that one never computes the inverse transform or
an expression or parameter that doesn't result from the corresponding
transform. If this would be the case then we don't need explicit functions for
the inverse transforms as we could save a pointer back to the untransformed
expression or parameter when transforming it. However, this is currently not
the case. When hash consing of expressions is enabled, situations arrise in
the interpreter where we take
  sensitivity-tranform-inverse
  unreverseify-parameter
  reverse-transform-inverse
of expressions and parameters that don't have back pointers. When hash consing
of expressions is disabled, situations arrise in the interpreter where we take
  perturbation-transform-inverse
  forward-transform-inverse
  sensitivity-tranform-inverse
  unreverseify-parameter
  reverse-transform-inverse
of expressions and parameters that don't have back pointers.

This version exhibits this phenomenon. I don't yet understand why.

To the best of my knowledge, this version runs both examples/run and
examples/forward/run correctly.